pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mango-evolve"
version = "0.1.0"
description = "LLM-driven evolutionary code generation for optimization problems"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "scipy>=1.10.0",
    "tenacity>=8.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-timeout>=2.0",
    "ruff>=0.8.0",
    "ty>=0.0.1a5",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict parameter
    "SIM108", # use ternary operator instead of if-else block
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG001", "ARG002"]  # unused arguments common in test fixtures
"experiments/*" = ["ARG001", "ARG002"]  # unused arguments in experimental code

[tool.ruff.lint.isort]
known-first-party = ["mango_evolve"]

[tool.ty.environment]
python-version = "3.10"
